refactor: apply code review feedback and update folder structure
