+++ week3/assignment1/src/App.jsx
@@ -1,27 +1,11 @@
-// App.jsx
-
-import React, { useState } from 'react';
-import Header from './components/Header';
-import Search from './components/Search';
-import Game from './components/Game';
+import { useState } from 'react';
+import Header from './components/header/Header';
+import GithubSearch from './pages/githubSearch/GithubSearch';
+import BaseballGame from './pages/baseballGame/BaseballGame';
 
 /** @jsxImportSource @emotion/react */
-import { Global, css } from '@emotion/react';
-
-const globalStyles = css`
-  html {
-    scroll-behavior: smooth;
-    font-size: 16px;
-  }
-  
-  body {
-    margin: 0;
-    padding: 0;
-    font-family: 'Arial', 'Apple SD Gothic Neo', 'Segoe UI', sans-serif;
-    background-color:rgb(245, 245, 245);
-    box-sizing: border-box;
-  }
-`;
+import { Global } from '@emotion/react';
+import { globalStyles } from './styles/globalStyles';
 
 function App() {
   const [activeTab, setActiveTab] = useState('search');
@@ -31,11 +15,11 @@ function App() {
       <Global styles={globalStyles} />
       <Header activeTab={activeTab} setActiveTab={setActiveTab} />
       <main>
-          {activeTab === 'search' && <Search />}
-          {activeTab === 'game' && <Game />}
+        {activeTab === 'search' && <GithubSearch />}
+        {activeTab === 'game' && <BaseballGame />}
       </main>
     </>
-  )
+  );
 }
 
-export default App
+export default App;
