+++ week3/assignment1/src/components/message/Message.styles.js
@@ -1,0 +1,7 @@
+import { css } from '@emotion/react';
+
+export const conditionalMsg = (theme) => css`
+  font-size: ${theme.fonts.md};
+  margin-top: 2rem;
+  font-weight: 500;
+`;
