+++ week3/assignment1/src/components/spinner/Spinner.jsx
@@ -1,0 +1,8 @@
+/** @jsxImportSource @emotion/react */
+import { spinner } from './Spinner.styles';
+
+const Spinner = () => {
+  return <div css={spinner}></div>;
+};
+
+export default Spinner;
