+++ week3/assignment1/src/main.jsx
@@ -1,5 +1,3 @@
-// main.jsx
-
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import { ThemeProvider } from '@emotion/react';
