+++ week3/assignment1/src/pages/githubSearch/GithubSearch.styles.js
@@ -1,0 +1,8 @@
+import { css } from '@emotion/react';
+
+export const container = css`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: flex-start;
+`;
