+++ week3/assignment1/src/styles/globalStyles.js
@@ -1,0 +1,16 @@
+import { css } from '@emotion/react';
+
+export const globalStyles = css`
+  html {
+    scroll-behavior: smooth;
+    font-size: 16px;
+  }
+
+  body {
+    margin: 0;
+    padding: 0;
+    font-family: 'Arial', 'Apple SD Gothic Neo', 'Segoe UI', sans-serif;
+    background-color: rgb(245, 245, 245);
+    box-sizing: border-box;
+  }
+`;
